requirements.txt
click>=8.0.0
mysql-connector-python>=8.0.0
rich>=10.0.0
python-dotenv>=1.0.0 